feat: continuous time simulation

+++ src/simulator/DefaultSimulator.h
@@ -15,7 +15,7 @@ public:
 
     DefaultSimulator(Graph &graph, std::vector<Agent> &agents, unsigned int seed) : Simulator(graph, agents, seed) {}
 
-    int simulate(unsigned int &currentTimestep, unsigned int maxTimeStep,
+    int simulate(double &currentTimestep, unsigned int maxTimeStep,
                  unsigned int delayStart = INT_MAX, unsigned int delayInterval = INT_MAX) override;
 
     void print(std::ostream &out) const override;
